requirements-dev.txt
astroid==2.9.1
attrs==21.4.0
awsebcli==3.20.2
bcrypt==3.2.0
blessed==1.19.0
botocore==1.21.65
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.3
cryptography==36.0.1
django-pytest==0.2.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
paramiko==2.9.1
pathspec==0.5.9
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pylint==2.12.2
PyNaCl==1.4.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.1.9
websocket-client==0.59.0
wrapt==1.13.3
